Listeners. Выносим handlePasteEvent в ClipboardManager. ClipboardManager переписываем на TS.
